+++ package.json
@@ -43,6 +43,7 @@
 		"@ai-sdk/google": "^1.2.19",
 		"@ai-sdk/openai": "^1.3.22",
 		"@ai-sdk/svelte": "^2.1.12",
+		"@clerk/themes": "^2.3.0",
 		"@datastax/astra-db-ts": "^1.5.0",
 		"@google/genai": "^0.7.0",
 		"@supabase/supabase-js": "^2.49.3",
@@ -50,7 +51,6 @@
 		"@uploadthing/svelte": "^7.3.0",
 		"@vercel/analytics": "^1.5.0",
 		"ai": "^4.3.16",
-		"aos": "^2.3.4",
 		"dompurify": "^3.2.4",
 		"lucide-svelte": "^0.485.0",
 		"marked": "^15.0.7",
