+++ src/app.css
@@ -7,7 +7,6 @@
 		light --default,
 		dark --prefersdark,
 		dracula,
-		retro,
 		nord;
 }
