+++ src/routes/mock.svelte
@@ -1,9 +1,13 @@
 <script lang="ts">
 	import { ArrowLeft, Eye, Flag } from 'lucide-svelte';
+
+	let showSolution = $state(false);
 </script>
 
 <div class="hidden lg:block size-2/5 lg:border-r border-base-300">
-	<button class="btn btn-ghost mt-4 ms-2"> <ArrowLeft />Back</button>
+	<button class="btn btn-ghost mt-4 ms-2">
+		<ArrowLeft />Back
+	</button>
 
 	<div class="mx-8 mt-4">
 		<p class="font-bold text-sm tracking-wide text-secondary mb-2">CHAPTER 2</p>
@@ -17,10 +21,12 @@
 			<div class="card-body">
 				<div class="flex flex-row justify-between border-b pb-2">
 					<h2 class="card-title">Solution</h2>
-					<button class="btn btn-ghost"><Eye /></button>
+					<button class="btn btn-ghost" onclick={() => (showSolution = !showSolution)}
+						><Eye /></button
+					>
 				</div>
 
-				<p class="blur mt-2">
+				<p class="{showSolution ? '' : 'blur'} mt-2">
 					Ophthalmic fluoroquinolones are used for severe infections, have increased affinity for
 					topoisomerase IV, and are safe for children.
 				</p>
@@ -33,10 +39,11 @@
 	<div class="lg:hidden flex flex-row mt-2 items-center justify-between">
 		<div class="flex-shrink-0 w-24"></div>
 	</div>
+
 	<div class="flex flex-row w-full lg:mt-6 mt-4 overflow-y-scroll">
 		<div class="indicator">
 			<button class="btn btn-circle btn-primary mx-2">1</button>
-			{#each Array(18) as _, i}
+			{#each Array(18), i (i)}
 				<button class="btn btn-circle btn-soft mx-2">{i + 2}</button>
 			{/each}
 		</div>
@@ -44,7 +51,7 @@
 
 	<div class="border-t border-base-300 w-full my-6"></div>
 
-	<div class=" mb-8 mt-2">
+	<div class="mb-8 mt-2">
 		<div class="mx-6 sm:mx-8">
 			<div class="font-bold text-lg mb-4">
 				Which of the following are true about ophthalmic fluoroquinolones?
@@ -57,26 +64,30 @@
 						<input type="checkbox" class="checkbox checkbox-primary" />
 					</div>
 				</label>
+
 				<label class="label cursor-pointer bg-base-200 rounded-full flex items-center">
-					<span class="flex-grow ml-8 my-2"
-						>B. They have increased affinity for topoisomerase IV</span
-					>
+					<span class="flex-grow ml-8 my-2">
+						B. They have increased affinity for topoisomerase IV
+					</span>
 					<div class="flex items-center justify-center w-16 mr-4">
 						<input type="checkbox" class="checkbox checkbox-primary" />
 					</div>
 				</label>
+
 				<label class="label cursor-pointer bg-base-200 rounded-full flex items-center">
 					<span class="flex-grow ml-8 my-2">C. They are used for viral infections</span>
 					<div class="flex items-center justify-center w-16 mr-4">
 						<input type="checkbox" class="checkbox checkbox-primary" />
 					</div>
 				</label>
+
 				<label class="label cursor-pointer bg-base-200 rounded-full flex items-center">
 					<span class="flex-grow ml-8 my-2">D. They are safe for children</span>
 					<div class="flex items-center justify-center w-16 mr-4">
 						<input type="checkbox" class="checkbox checkbox-primary" />
 					</div>
 				</label>
+
 				<label class="label cursor-pointer bg-base-200 rounded-full flex items-center">
 					<span class="flex-grow ml-8 my-2">E. They are used for fungal infections</span>
 					<div class="flex items-center justify-center w-16 mr-4">
@@ -85,9 +96,9 @@
 				</label>
 			</div>
 		</div>
+
 		<div class="flex flex-row justify-center mt-8 gap-4">
 			<button class="btn btn-outline">Clear</button>
-
 			<button class="btn btn-outline btn-success">Check</button>
 			<button class="btn btn-warning btn-outline">
 				<Flag />
